make monolineedit private
